refactor: colsole.log 삭제
