+++ client/src/components/Cards/storeCard.style.ts
