+++ client/src/pages/store/StoreList.tsx
@@ -1,6 +1,6 @@
 import { FaSearch } from 'react-icons/fa';
 import { useState, useEffect, useRef, ChangeEvent, KeyboardEvent } from 'react';
-import StoreCard from '../../components/storeCard';
+import StoreCard from '../../components/Cards/storeCard';
 import axiosInstance from '../../api/apis';
 import { Store, PageInfo } from '../../assets/interface/Store.interface';
 import { 
@@ -29,7 +29,6 @@ function StoreList() {
     rootMargin: '0px',
     threshold: 1.0
   };
-  console.log(filteredStores)
   useEffect(() => {
     const handleObserver = (entries: IntersectionObserverEntry[]) => {
       entries.forEach((entry) => {
